REF: Create separate module for architecture-test
