+++ android/library/architecture-test/.gitignore
@@ -1,0 +1,1 @@
+/build
